Made writers static, added item writer, improved skill penalty item values

+++ Ambermoon.Data.Legacy/Serialization/ItemSlotWriter.cs
@@ -8,7 +8,7 @@ namespace Ambermoon.Data.Legacy.Serialization
         {
             dataWriter.Write((byte)itemSlot.Amount);
             dataWriter.Write((byte)itemSlot.NumRemainingCharges);
-            dataWriter.Write((byte)itemSlot.RechargeTimes);
+            dataWriter.Write(itemSlot.RechargeTimes);
             dataWriter.WriteEnumAsByte(itemSlot.Flags);
             dataWriter.Write((ushort)itemSlot.ItemIndex);
         }
